dsound: Fix rolloff factor calculation.
We should treat the rolloff factor as a factor of the distance of the
buffer. For example, a rolloff factor of 2 should treat the buffer as if
it were twice as far away from the listener as it is.

The previous calculation would instead double the decibels of
attenuation.

Signed-off-by: Andrew Eikum <aeikum@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>
(cherry picked from commit 1671a83f55bee46ee57cc0c8869e943d67f14ea5)
Signed-off-by: Michael Stefaniuc <mstefani@winehq.org>